# src/main/java/com/reclamos/backend/dto/request/AssignLabelsRequest.java
package com.reclamos.backend.dto.request;

import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.Set;
import java.util.UUID;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class AssignLabelsRequest {

    @NotEmpty
    private Set<@NotNull UUID> labelIds;
}