# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: M2 - Atención Ciudadana API
  description: API REST del módulo M2 de Atención Ciudadana, Reclamos y Solicitudes.
  version: 1.0.0
servers:
  - url: /
    description: Mismo origen que sirve Swagger UI
...
paths:
...
  /api/tickets/{ticketId}:
    get:
      tags: [ Tickets ]
      summary: Consultar el detalle de un ticket propio
      description: |
        GET /tickets/{id} (Entidades V1.49 §"Autenticación, roles y vistas"):
        detalle ciudadano de un ticket, sólo para su propietario autenticado.
        Un ticket anónimo no tiene citizenId, así que nunca es "propio" de
        nadie acá; su consulta pasa por POST /tracking/access.
      operationId: getTicketById
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Detalle del ticket propio.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetailResponse'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/staff/labels:
    get:
      tags: [ Labels ]
      summary: Listar el catálogo completo de etiquetas
      operationId: listLabels
      security: [ { bearerAuth: [ ] } ]
      responses:
        '200':
          description: Etiquetas ordenadas por nombre e id.
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/LabelResponse' } } } }
        '401': { $ref: '#/components/responses/InvalidToken' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [ Labels ]
      summary: Crear una etiqueta
      operationId: createLabel
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateLabelRequest' } } }
      responses:
        '201': { description: Etiqueta creada, content: { application/json: { schema: { $ref: '#/components/schemas/LabelResponse' } } } }
        '409': { description: Código duplicado }
  /api/staff/labels/{labelId}:
    get:
      tags: [ Labels ]
      operationId: getLabel
      security: [ { bearerAuth: [ ] } ]
      parameters: [ { name: labelId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '200': { description: Etiqueta, content: { application/json: { schema: { $ref: '#/components/schemas/LabelResponse' } } } }, '404': { $ref: '#/components/responses/NotFound' } }
    put:
      tags: [ Labels ]
      operationId: updateLabel
      security: [ { bearerAuth: [ ] } ]
      parameters: [ { name: labelId, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateLabelRequest' } } } }
      responses: { '200': { description: Etiqueta modificada, content: { application/json: { schema: { $ref: '#/components/schemas/LabelResponse' } } } } }
    delete:
      tags: [ Labels ]
      operationId: deleteLabel
      security: [ { bearerAuth: [ ] } ]
      parameters: [ { name: labelId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '204': { description: Eliminada }, '409': { description: La etiqueta está referenciada } }
  /api/staff/tickets/{ticketId}/labels:
    post:
      tags: [ Labels ]
      summary: Asignar etiquetas manuales de manera atómica e idempotente
      operationId: assignTicketLabels
      security: [ { bearerAuth: [ ] } ]
      parameters: [ { $ref: '#/components/parameters/TicketId' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AssignLabelsRequest' } } } }
      responses: { '204': { description: Asignadas }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
  /api/staff/tickets/{ticketId}/labels/{labelId}:
    delete:
      tags: [ Labels ]
      summary: Retirar una etiqueta de forma idempotente
      operationId: removeTicketLabel
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { $ref: '#/components/parameters/TicketId' }
        - { name: labelId, in: path, required: true, schema: { type: string, format: uuid } }
      responses: { '204': { description: Retirada o ya ausente }, '403': { $ref: '#/components/responses/Forbidden' } }
  /api/staff/tickets/{ticketId}:
    get:
      tags: [ Tickets ]
      summary: Consultar el detalle staff de un ticket
      description: |
        GET /staff/tickets/{id} (Guía funcional M2 §7/§7.1): detalle de
        gestión interna. AGENT y ADMIN acceden a cualquier ticket ajeno sin
        restricción de área; AREA_RESPONSIBLE sólo a los de su propia areaId
        (o al suyo propio como ciudadano). El endpoint citizen-view expone
        en cambio la misma proyección pública que recibe el propietario.
      operationId: getStaffTicketDetail
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Detalle staff del ticket.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetailResponse'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/staff/tickets/{ticketId}/citizen-view:
    get:
      tags: [ Tickets ]
      summary: Consultar un ticket ajeno con la proyección ciudadana, desde el contexto staff
      description: |
        GET /staff/tickets/{id}/citizen-view (Guía funcional M2 §7.1): misma
        proyección visible al ciudadano, pero consultada desde el contexto
        staff — no cambia la identidad de quien pregunta ni suplanta al
        propietario. Mismo control de acceso que
        GET /staff/tickets/{id}: AGENT y ADMIN acceden a cualquier ticket
        ajeno; AREA_RESPONSIBLE sólo a los de su propia areaId (o al suyo
        propio como ciudadano).
      operationId: getStaffTicketCitizenView
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Proyección ciudadana del ticket.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetailResponse'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/staff/tickets/{ticketId}/duplicate-candidates:
    get:
      tags: [ Tickets ]
      summary: Sugerir posibles tickets duplicados
      description: |
        Busca tickets activos de la misma categoría dentro del radio geográfico
        y de la ventana temporal configurados. Sólo AGENT y ADMIN pueden usar
        esta herramienta de triage. La operación es de sólo lectura: no vincula
        tickets ni modifica sus estados. Una lista vacía es una respuesta válida
        e indica que no se encontraron candidatos.
      operationId: getDuplicateCandidates
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Sugerencias ordenadas por distancia y diferencia temporal; puede ser una lista vacía.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DuplicateCandidateResponse'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: El ticket origen no existe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token Bearer obtenido del proveedor de identidad vigente. En DEV puede obtenerse mediante /api/auth/login.
  parameters:
...
    TicketId:
      name: ticketId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    InvalidToken:
      description: Token Bearer ausente o inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: INVALID_TOKEN
            message: La sesión no es válida
    Forbidden:
      description: La identidad autenticada no tiene permiso para realizar la operación.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: FORBIDDEN
            message: No está autorizado para realizar esta operación
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: NOT_FOUND
            message: Recurso no encontrado
...
  schemas:
    TicketPublicId:
      type: string
      description: Identificador humano estable del ticket; no reemplaza al UUID ni al código secreto de tracking.
      pattern: '^TK-[0-9]{4}-[0-9]{6,}$'
      example: TK-2026-000123
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
      additionalProperties: false
...
    CreateLabelRequest:
      type: object
      required: [ code, name ]
      properties: { code: { type: string, maxLength: 100 }, name: { type: string, maxLength: 150 }, description: { type: string, nullable: true } }
    UpdateLabelRequest:
      type: object
      required: [ name, active ]
      properties: { name: { type: string, maxLength: 150 }, description: { type: string, nullable: true }, active: { type: boolean } }
    AssignLabelsRequest:
      type: object
      required: [ labelIds ]
      properties: { labelIds: { type: array, minItems: 1, uniqueItems: true, items: { type: string, format: uuid } } }
    LabelResponse:
      type: object
      required: [ id, code, name, description, active ]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        active:
          type: boolean
    LabelSummaryResponse:
      type: object
      required: [ id, code, name, description ]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
    TicketDetailResponse:
      type: object
      description: |
        Detalle plano del ticket. Conserva los campos útiles de TicketResponse,
        agrega descripción, metadatos de adjuntos e historial y expone sólo el
        nombre del barrio. En detalle ciudadano y citizen-view los adjuntos son
        PUBLIC y las actividades usan la proyección pública sanitizada; el detalle
        staff puede incluir adjuntos INTERNAL y campos adicionales del historial.
      required: [id, publicId, requestTypeCode, requestTypeName, categoryName, subcategoryName,
                 ticketType, summary, description, currentStatus, currentPriority, responsibleAreaId,
                 assignedAgentId, anonymous, estimatedAffectedCount, escalated, escalationReasonCode,
                 escalatedAt, firstResponseDueAt, firstResponseNearDue, firstResponseBreached,
                 resolutionDueAt, slaNearDue, slaBreached, resolutionNearDueAt, neighborhoodName,
                 classificationFinalizedAt, statusChangedAt, createdAt, updatedAt, attachments,
                 ticketActivities, labels]
      properties:
        id:
          type: string
          format: uuid
        publicId:
          $ref: '#/components/schemas/TicketPublicId'
        requestTypeCode:
          type: string
        requestTypeName:
          type: string
        categoryName:
          type: string
        subcategoryName:
          type: string
        ticketType:
          $ref: '#/components/schemas/TicketType'
        summary:
          type: string
        description:
          type: string
        currentStatus:
          $ref: '#/components/schemas/TicketStatus'
        currentPriority:
          $ref: '#/components/schemas/Priority'
        responsibleAreaId:
          type: string
        assignedAgentId:
          type: string
          nullable: true
        anonymous:
          type: boolean
        estimatedAffectedCount:
          type: integer
          format: int32
        escalated:
          type: boolean
        escalationReasonCode:
          allOf:
            - $ref: '#/components/schemas/EscalationReasonCode'
          nullable: true
        escalatedAt:
          type: string
          format: date-time
          nullable: true
        firstResponseDueAt:
          type: string
          format: date-time
          nullable: true
        firstResponseNearDue:
          type: boolean
        firstResponseBreached:
          type: boolean
        resolutionDueAt:
          type: string
          format: date-time
          nullable: true
        slaNearDue:
          type: boolean
        slaBreached:
          type: boolean
        resolutionNearDueAt:
          type: string
          format: date-time
          nullable: true
        neighborhoodName:
          type: string
          nullable: true
        classificationFinalizedAt:
          type: string
          format: date-time
          nullable: true
        statusChangedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        attachments:
          type: array
          description: Metadatos únicamente; este contrato no provee descarga del archivo.
          items:
            $ref: '#/components/schemas/TicketAttachmentResponse'
        ticketActivities:
          type: array
          description: Historial ordenado por sequence ascendente.
          items:
            $ref: '#/components/schemas/TicketActivityResponse'
        labels:
          type: array
          items:
            $ref: '#/components/schemas/LabelSummaryResponse'
    TicketAttachmentResponse:
      type: object
      description: Metadatos seguros de un adjunto, sin claves ni internals de almacenamiento.
      required: [id, fileName, contentType, sizeBytes, visibility, createdAt]
      properties:
        id:
          type: integer
          format: int64
        fileName:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        visibility:
          type: string
          enum: [PUBLIC, INTERNAL]
          description: Citizen y citizen-view reciben sólo PUBLIC; staff puede recibir ambas.
        createdAt:
          type: string
          format: date-time
    TicketActivityResponse:
      type: object
      description: |
        Actividad sanitizada. actorType, prioridades y message son nulos en la
        proyección ciudadana; staff puede recibirlos. Nunca expone identificadores
        de actor/módulo/evento, metadata cruda ni ticketVersion.
      required: [sequence, actionType, previousStatus, newStatus, occurredAt, reasonCode,
                 actorType, previousPriority, newPriority, message]
      properties:
        sequence:
          type: integer
          format: int32
        actionType:
          type: string
          enum: [TICKET_CREATED, REVIEW_STARTED, STATE_CHANGED, REQUEST_TYPE_CHANGED, ROUTED,
                 RETURNED_BY_AREA, PRIORITY_CHANGED, SLA_NEAR_DUE, SLA_BREACHED, ESCALATED,
                 INFORMATION_REQUIRED, INFORMATION_PROVIDED, PROGRESS_REPORTED, DUPLICATE_LINKED,
                 RESOLVED, REOPENED, CANCELLATION_REQUESTED, CANCELLATION_APPROVED,
                 CANCELLATION_REJECTED, CANCELLED, CLOSED, PUBLIC_MESSAGE_SENT,
                 INTERNAL_MESSAGE_ADDED, ATTACHMENT_ADDED]
        previousStatus:
          allOf:
            - $ref: '#/components/schemas/TicketStatus'
          nullable: true
        newStatus:
          allOf:
            - $ref: '#/components/schemas/TicketStatus'
          nullable: true
        occurredAt:
          type: string
          format: date-time
        reasonCode:
          type: string
          nullable: true
        actorType:
          type: string
          enum: [CITIZEN, AGENT, AREA_RESPONSIBLE, ADMIN, EXTERNAL_USER, SYSTEM]
          nullable: true
        previousPriority:
          allOf:
            - $ref: '#/components/schemas/Priority'
          nullable: true
        newPriority:
          allOf:
            - $ref: '#/components/schemas/Priority'
          nullable: true
        message:
          type: string
          nullable: true
    DuplicateCandidateResponse:
      type: object
      description: Sugerencia explicable de un posible duplicado; no representa una vinculación automática.
      required: [ ticketId, publicId, summary, currentStatus, categoryId, categoryName,
                  requestTypeCode, requestTypeName, neighborhoodId, neighborhoodName,
                  distanceMeters, timeDifferenceMinutes, createdAt ]
      properties:
        ticketId:
          type: string
          format: uuid
        publicId:
          $ref: '#/components/schemas/TicketPublicId'
        summary:
          type: string
        currentStatus:
          $ref: '#/components/schemas/TicketStatus'
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        requestTypeCode:
          type: string
        requestTypeName:
          type: string
        neighborhoodId:
          type: string
          format: uuid
          nullable: true
        neighborhoodName:
          type: string
          nullable: true
        distanceMeters:
          type: number
          format: double
          minimum: 0
        timeDifferenceMinutes:
          type: integer
          format: int64
          minimum: 0
        createdAt:
          type: string
          format: date-time
...
    TicketType:
      type: string
      enum: [COMPLAINT, REQUEST, INQUIRY, SUGGESTION]
    TicketStatus:
      type: string
      enum: [REGISTERED, IN_REVIEW, ROUTED, IN_PROGRESS, PENDING_INFORMATION, DUPLICATE, RESOLVED, CLOSED, CANCELLED]
    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
...
    EscalationReasonCode:
      type: string
      enum: [CRITICAL_PRIORITY, SLA_BREACHED, MANUAL]
...